tests/submission/Microsoft/apis/getElementsByClassName_012.htm

<!DOCTYPE html>
<html><head class="foo">
        <title class="foo">getElementsByClassName</title>
        <meta class="foo" content="big element listing" name="description">
        <link class="foo">
        <base class="foo">
        <script class="foo"></script>
        <style class="foo"></style>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link href="http://www.w3.org/TR/domcore/#dom-document-getelementsbyclassname" rel="help">
</head>
    <body class="foo">
    <div id='log'></div>
        <a class="foo">a</a>
        <abbr class="foo">abbr</abbr>
        <acronym class="foo">acronym</acronym>
        <address class="foo">address</address>
        <applet class="foo">applet</applet>
        <b class="foo">b</b>
        <bdo class="foo">bdo</bdo>
        <big class="foo">big</big>
        <blockquote class="foo">blockquote</blockquote>
        <br class="foo">
        <button class="foo">button</button>
        <center class="foo">center</center>
        <cite class="foo">cite</cite>
        <code class="foo">code</code>
        <del class="foo">del</del>
        <dfn class="foo">dfn</dfn>
        <dir class="foo">dir
            <li class="foo">li</li>
        </dir>
        <div class="foo">div</div>
        <dl class="foo">
            <dt class="foo">
                </dt><dd class="foo">dd</dd>
            
        </dl>
        <em class="foo">em</em>
        <font class="foo">font</font>
        <form class="foo">
            <label class="foo">label</label>
            <fieldset class="foo">
                <legend class="foo">legend</legend>
            </fieldset>
        </form>
        <h1 class="foo">h1</h1>
        <hr class="foo">
        <i class="foo">i</i>
        <iframe class="foo">iframe</iframe>
        <img class="foo">
        <input class="foo">
        <ins class="foo">ins</ins>
        <kbd class="foo">kbd</kbd>
        <map class="foo">
            <area class="foo"></area>
        </map>                                  
        <menu class="foo">menu</menu>
        <noscript class="foo">noscript</noscript>
        <object class="foo">
            <param class="foo">
        </object>
        <ol class="foo">ol</ol>
        <p class="foo">p</p>
        <pre class="foo">pre</pre>
        <q class="foo">q</q>
        <s class="foo">s</s>
        <samp class="foo">samp</samp>
        <select class="foo">
            <optgroup class="foo">optgroup</optgroup>
            <option class="foo">option</option>
        </select>
        <small class="foo">small</small>
        <span class="foo">span</span>
        <strike class="foo">strike</strike>
        <strong class="foo">strong</strong>
        <sub class="foo">sub</sub>
        <sup class="foo">sup</sup>
        colgroup<table class="foo">
            <caption class="foo">caption</caption>
            <colgroup><col class="foo">
            </colgroup><colgroup class="foo"></colgroup>
            <thead class="foo">
                <tr><th class="foo">th</th>
            </tr></thead>
            <tbody class="foo">
                <tr class="foo">
                    <td class="foo">td</td>
                </tr>
            </tbody>
            <tfoot class="foo"></tfoot>
        </table>
        <textarea class="foo">textarea</textarea>
        <tt class="foo">tt</tt>
        <u class="foo">u</u>
        <ul class="foo">ul</ul>
        <var class="foo">var</var>
        <script type="text/javascript">
 	  test(function ()
          {
           var arrElements = new Array();
                arrElements[0] = "HEAD";
                arrElements[1] = "TITLE";
                arrElements[2] = "META";
                arrElements[3] = "LINK";
                arrElements[4] = "BASE";
                arrElements[5] = "SCRIPT";
                arrElements[6] = "STYLE";
                arrElements[7] = "BODY";
                arrElements[8] = "A";
                arrElements[9] = "ABBR";
                arrElements[10] = "ACRONYM";
                arrElements[11] = "ADDRESS";
                arrElements[12] = "APPLET";
                arrElements[13] = "B";
                arrElements[14] = "BDO";
                arrElements[15] = "BIG";
                arrElements[16] = "BLOCKQUOTE";
                arrElements[17] = "BR";
                arrElements[18] = "BUTTON";
                arrElements[19] = "CENTER";
                arrElements[20] = "CITE";
                arrElements[21] = "CODE";
                arrElements[22] = "DEL";
                arrElements[23] = "DFN";
                arrElements[24] = "DIR";
                arrElements[25] = "LI";
                arrElements[26] = "DIV";
                arrElements[27] = "DL";
                arrElements[28] = "DT";
                arrElements[29] = "DD";
                arrElements[30] = "EM";
                arrElements[31] = "FONT";
                arrElements[32] = "FORM";
                arrElements[33] = "LABEL";
                arrElements[34] = "FIELDSET";
                arrElements[35] = "LEGEND";
                arrElements[36] = "H1";
                arrElements[37] = "HR";
                arrElements[38] = "I";
                arrElements[39] = "IFRAME";
                arrElements[40] = "IMG";
                arrElements[41] = "INPUT";
                arrElements[42] = "INS";
                arrElements[43] = "KBD";
                arrElements[44] = "MAP";
                arrElements[45] = "AREA";
                arrElements[46] = "MENU";
                arrElements[47] = "NOSCRIPT";
                arrElements[48] = "OBJECT";
                arrElements[49] = "PARAM";
                arrElements[50] = "OL";
                arrElements[51] = "P";
                arrElements[52] = "PRE";
                arrElements[53] = "Q";
                arrElements[54] = "S";
                arrElements[55] = "SAMP";
                arrElements[56] = "SELECT";
                arrElements[57] = "OPTGROUP";
                arrElements[58] = "OPTION";
                arrElements[59] = "SMALL";
                arrElements[60] = "SPAN";
                arrElements[61] = "STRIKE";
                arrElements[62] = "STRONG";
                arrElements[63] = "SUB";
                arrElements[64] = "SUP";
                arrElements[65] = "TABLE";
                arrElements[66] = "CAPTION";
                arrElements[67] = "COL";
                arrElements[68] = "COLGROUP";
                arrElements[69] = "THEAD";
                arrElements[70] = "TH";
                arrElements[71] = "TBODY";
                arrElements[72] = "TR";
                arrElements[73] = "TD";
                arrElements[74] = "TFOOT";
                arrElements[75] = "TEXTAREA";
                arrElements[76] = "TT";
                arrElements[77] = "U";
                arrElements[78] = "UL";
                arrElements[79] = "VAR";
            var bResult = true;

                var collection = document.getElementsByClassName("foo");
                
                if (collection.length == 80)
                {
                    for (var x = 0; x < collection.length; x++)
                    {
                        assert_equals(collection[x].nodeName, arrElements[x]);
                    }
                }
}, "big element listing");
        </script>
    
</body></html>
